# Conflicts:
#	features/user/chatbot/build.gradle
#	features/user/chatbot/src/main/java/com/tokopedia/chatbot/ChatbotConstant.kt
#	features/user/chatbot/src/main/java/com/tokopedia/chatbot/di/ChatbotComponent.kt
#	features/user/chatbot/src/main/java/com/tokopedia/chatbot/di/ChatbotModule.kt
#	features/user/chatbot/src/main/java/com/tokopedia/chatbot/view/fragment/ChatbotFragment.kt

+++ features/merchant/statistic/src/main/java/com/tokopedia/statistic/analytics/StatisticTracker.kt
@@ -165,7 +165,7 @@ object StatisticTracker {
         TrackingHelper.sendGeneralEvent(map)
     }
 
-    fun sendImpressionProgressBarEvent(dataKey: String, stateColor: String, valueScore: Int) {
+    fun sendImpressionProgressBarEvent(dataKey: String, stateColor: String, valueScore: Long) {
         val map = TrackingHelper.createMap(
             TrackingConstant.VIEW_SELLER_WIDGET,
             arrayOf(TrackingConstant.SELLER_APP, TrackingConstant.SHOP_INSIGHT).joinToString(" - "),
@@ -178,7 +178,7 @@ object StatisticTracker {
         TrackingHelper.sendGeneralEvent(map)
     }
 
-    fun sendClickProgressBarEvent(dataKey: String, stateColor: String, valueScore: Int) {
+    fun sendClickProgressBarEvent(dataKey: String, stateColor: String, valueScore: Long) {
         val map = TrackingHelper.createMap(
             TrackingConstant.CLICK_SHOP_INSIGHT,
             arrayOf(TrackingConstant.SELLER_APP, TrackingConstant.SHOP_INSIGHT).joinToString(" - "),
